fix(Filter): change add options logic cause it had bugs.
